Move symmetry code
