+++ src/filters/mod.rs
@@ -6,6 +6,7 @@ pub mod median;
 pub mod min_max;
 pub mod prewitt;
 pub mod sobel;
+pub mod symmetry;
 
 // TODO We might want to offer all NumPy mode (use PadMode instead)
 /// Method that will be used to determines how the input array is extended beyond its boundaries.
